add additional packages required by fastai (#435)
* add additional packages required by fastai
* add unit test for fastai
