+++ tests/smoke/test_notebooks_gpu.py
@@ -33,3 +33,27 @@ def test_ncf_smoke(notebooks):
     assert results["ndcg"] == pytest.approx(0.193496, TOL)
     assert results["precision"] == pytest.approx(0.175504, TOL)
     assert results["recall"] == pytest.approx(0.100301, TOL)
+
+@pytest.mark.smoke
+@pytest.mark.gpu
+def test_fastai(notebooks):
+    assert get_number_gpus() >= 1
+
+    notebook_path = notebooks["fastai"]
+    pm.execute_notebook(notebook_path, OUTPUT_NOTEBOOK, kernel_name=KERNEL_NAME)
+    pm.execute_notebook(
+        notebook_path,
+        OUTPUT_NOTEBOOK,
+        kernel_name=KERNEL_NAME,
+        parameters=dict(TOP_K=10, MOVIELENS_DATA_SIZE="100k"),
+    )
+    results = pm.read_notebook(OUTPUT_NOTEBOOK).dataframe.set_index("name")["value"]
+
+    assert results["rmse"] == pytest.approx(0.912115, TOL)
+    assert results["mae"] == pytest.approx(0.723051, TOL)
+    assert results["rsquared"] == pytest.approx(0.356302, TOL)
+    assert results["exp_var"] == pytest.approx(0.357081, TOL)
+    assert results["map"] == pytest.approx(0.021485, TOL)
+    assert results["ndcg"] == pytest.approx(0.137494, TOL)
+    assert results["precision"] == pytest.approx(0.124284, TOL)
+    assert results["recall"] == pytest.approx(0.045587, TOL)
